+++ PKHeX.Core/Legality/Areas/EncounterArea3.cs
@@ -66,11 +66,11 @@ namespace PKHeX.Core
 
         private EncounterSlot3 ReadRegularSlot(ReadOnlySpan<byte> entry)
         {
-            int species = ReadUInt16LittleEndian(entry);
-            int form = entry[2];
+            ushort species = ReadUInt16LittleEndian(entry);
+            byte form = entry[2];
             byte slotNum = entry[3];
-            int min = entry[4];
-            int max = entry[5];
+            byte min = entry[4];
+            byte max = entry[5];
 
             byte mpi = entry[6];
             byte mpc = entry[7];
@@ -96,11 +96,11 @@ namespace PKHeX.Core
 
         private EncounterSlot3Swarm ReadSwarmSlot(ReadOnlySpan<byte> entry)
         {
-            int species = ReadUInt16LittleEndian(entry);
+            ushort species = ReadUInt16LittleEndian(entry);
             // form always 0
             byte slotNum = entry[3];
-            int min = entry[4];
-            int max = entry[5];
+            byte min = entry[4];
+            byte max = entry[5];
 
             int[] moves =
             {
