+++ PKHeX.Core/Legality/Areas/EncounterArea8a.cs
@@ -79,8 +79,8 @@ public sealed record EncounterArea8a : EncounterArea
             var entry = areaData.Slice(ofs, bpe);
             byte flawless = entry[7];
             var gender = (Gender)entry[6];
-            int max = entry[5];
-            int min = entry[4];
+            byte max = entry[5];
+            byte min = entry[4];
             var alpha = entry[3];
             var form = entry[2];
             var species = ReadUInt16LittleEndian(entry);
