+++ PKHeX.Core/Legality/Encounters/EncounterSlot/EncounterSlot5.cs
@@ -8,7 +8,7 @@ namespace PKHeX.Core
     {
         public override int Generation => 5;
 
-        public EncounterSlot5(EncounterArea5 area, int species, int form, int min, int max) : base(area, species, form, min, max)
+        public EncounterSlot5(EncounterArea5 area, ushort species, byte form, byte min, byte max) : base(area, species, form, min, max)
         {
         }
