+++ PKHeX.Core/Legality/Encounters/EncounterStatic/EncounterStatic3.cs
@@ -12,7 +12,7 @@ namespace PKHeX.Core
         public override int Generation => 3;
         public bool Roaming { get; init; }
 
-        public EncounterStatic3(int species, int level, GameVersion game) : base(game)
+        public EncounterStatic3(ushort species, byte level, GameVersion game) : base(game)
         {
             Species = species;
             Level = level;
